Major changes to debug mechanism
What we used to do to enable debug logging is this: give the "-f debug" flag to
cabal when installing snap-core, which used the C preprocessor to select a
compile-time implementation for the "debug" function.

What we're doing now instead is using "unsafePerformIO" to test the "DEBUG"
environment variable the first time you call the debug function; subsequent
calls to "debug" will re-use the fully-evaluated thunk, so you shouldn't pay a
performance penalty if debug logging is turned off. We do pay a slight penalty
because we can no longer inline calls to "debug", so debug overhead when
debugging is off goes from a no-op to a call into a function pointer;
benchmarks say the price to pay is slight and the productivity gains are
enormous: set "DEBUG=1" when you invoke your server to get debugging output
without recompiling the entire snap library stack.