Clean up uploaded temp files even if the source enumerator dies.
If the request body enumerator fails (for instance, the host drops the
connection), the temporary files were not being deleted because the iteratee's
exception handler never got a chance to run.

Fix this by attaching a finalizer to an MVar which will cause the files to be
cleaned up if this happens, even if the garbage collector has to do it
"sometime later".