Implement handleFormUploads
A high level handleFormUploads is intentended to be used with storage
functions like storeAsLazyByteString or withTemporaryStore.

In addition to UploadPolicy can be configured with FileUploadPolicy.

FileUploadPolicy is used to setup maximum uploaded file size and maximum
number of uploaded files. Also it can be used to set how files without
file names should be handled.

The setSkipFilesWithoutNames is useful to use with html5 forms.
According to html5 spec, form inputs with type file if not set are
encoded as files without name, with empty body and mime type of
application/octet-stream. Default FileUploadPolicy allows empty files
with zero length, it should fit this use-case. Note that older browsers
may omit such fields from encoded form-data, so this behaviour is
desireable.

If you use setStoreFilesWithoutNames then files will be passed to store
function which may decide how to deal with them, since it gets PartInfo
as an argument.